requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
pyjwt>=2.6.0
# Les dépendances ci-dessous seront ajoutées plus tard pour l'implémentation complète
# pinecone-client>=2.2.2
# langchain>=0.0.267
# langchain-pinecone>=0.0.1
# langchain-openai>=0.0.2
# langchain-text-splitters>=0.0.1
# openai>=0.27.8
# pdfplumber>=0.10.0
# python-docx>=0.8.11
# python-pptx>=0.6.21
# odfpy>=1.4.1
# PyPDF2>=3.0.1
# textract==1.6.4 